Fully support IPv6; match IPv4 & IPv6 separately.
Previously, `hosts` would update IPv4-addressed hostname & aliases with
IPv6 addresses, and vice versa. That's a bug. A hosts file allows for a
hostname to list both IPv4 and IPv6 addresses.

On the code itself, `parse-and-merge` encapsulates the entire state of
addresses, hostnames, and aliases. Using a procedural approach here is
relatively tedious and would benefit from an object model. That's an
exercise for a later patch to the `hosts` project.